Improve stability of HTMLTableElement.extract
Previously, the method required a table's column headers to be wrapped
inside a thead element; which, of course, was causing the method to
break for tables whose headers were erroneously included in the tbody
instead. The method now falls back on using the first row of the first
tbody tag if a thead wasn't found inside the table.