Change filenames of minified resources
"Main" is a misleading choice for a minified file's name when one of the
source files it's generated from is also named "main". For example, it's
possible for a reader to misconstrue "min/main.css" as a compressed form
of "css/main.css" alone, not taking into account the preceding two files
prepended to it ("css/fonts.css" & "css/global.css").