Integrate word-wrapping method
The class was still relying on an externally-defined helper function for
wrapping lines of text. The function extended the global String object's
prototype, which might poise compatibility issues in some environments.
Obviously this just won't do, so I've moved it to a method of the Record
Jar class instead.

Originally applied in Alhadis/Record-Jar@2b2f724d80