Allow graceful stubbing of Google Analytics
Assigning a noop to window.ga is a better way of dealing with unsupplied
tracking IDs, since GA won't overwrite window.ga if it already exists.

This has made it harder to stub event-tracking during development, as it
becomes easy to forget about that "var ga = window.ga || function(){ };"
line that's sitting at the top of a document when you get to deployment.