Add class to display a file's "human-readable format"
Probably the oldest surviving snippet of code in my filesystem, this was
written in late-ish 2009 while finishing my multimedia diploma. Seems as
though it was cleaned up at some point in July 2012 (file's modification
timestamp is 1343461428), which would explain why the formatting is less
shit than I expected it to be.

I'm archiving this snippet purely for historical curiosity, and possibly
for future laughs (which might also be accompanied by unhealthy doses of
depressing memories).