Ensure final line of STDIN gets trimmed correctly
When reading from standard input, the trim utility was ignoring the last
line if it didn't end with a trailing newline. This issue went unnoticed
because a complementary tool was appending newlines to its output before
piping it through to trim. When these superfluous line-breaks were fixed
in c69a8dde, the buggy handling of STDIN's final line became evident.

This behaviour was also noticed earlier this year in April (see bug #9),
but the solution applied in 99cd6517 only patched the program's trimming
of a specified file's content (when the -f flag was passed).

Closes #16.