Have prep-repo.sh create a new Git repo if needed
Since we're going to the effort of writing crude shortcuts to accelerate
workflow (see ed4b4d155), we may as well have it start the engine before
hitting the pedal.

This commit instructs prep-repo.sh to initialise a new Git repo if one's
not found in the current working directory. This behaviour is overridden
with the -u flag.