Extend Rainbow program with new options
The xterm-colour viewer now accepts the following options:

  -t  Declares a custom string to use instead of the alphabet

  -b  Fills each row with a solid-coloured bar using codepoint U+2588.
      Overrides the value of the -t option, if passed previously.

The program also checks the terminal's capabilities before attempting to
print anything. An error message will be printed to STDERR if running in
an environment that lacks support for coloured output.

The terminal's reported colour range will be used to limit values passed
to the -c option. For example, it's now impossible to request a range of
colours greater than 256 when running on an xterm-256color terminal.