Add rule for conditional WebP redirection
If a WebP-compatible browser requests an image that's been saved in WebP
format, redirect them to use that instead. Be aware this rule might have
unpredictable behaviour if two different images are saved in two formats
(each with the same filename). E.g.:

   * src/img/1.png    # Massive 24-bit transparent RGBA PNG
   * src/img/1.webp   # Highly-compressed version of the above
   * src/img/1.jpg    # Completely unrelated image, also saved as "1"

In the above hypothetical scenario, loading 1.jpg redirects a browser to
1.webp, displaying an entirely different image.