Remove ~ combinators from :checked hack
A strange bug in Chrome on iOS prevents selectors with ":checked ~" from
being applied. This issue only appears to affect the iPhone/iPad version
of Chrome; other devices are unaffected.

In the meantime, a workaround for this edge case is to use + combinators
for applying :checked-controlled styling. For example, wrap the affected
elements in a single container, then target them as descendants:

    #toggle-menu:checked + div > nav {
        max-height: none;
    }

    <input type="checkbox" id="toggle-menu" />
    <div>
        <label for="toggle-menu"> ☰ </label>
        <nav></nav>
    </div>

Isolated test case: http://s.codepen.io/Alhadis/debug/VvdKVr