Add bookmarklet to kill FB's registration nag
Facebook now displays a registration screen to unregistered users asking
them to sign up while browsing public pages. Since this is a pain in the
arse for managing band submissions at Metal Archives, I've written a toy
to kick Facebook in the gonads. So to speak.

Copy the last line of the script to a new browser bookmarklet, and click
it whenever the soulless, corporate, parasitic behemoth asks you to give
yourself up to it.

Fuck off.