Add tool to optimise SVG icons exported by Illustrator
This script deletes unnecessary junk added by Adobe Illustrator to files
saved in SVG format. Specifically, deprecated and superfluous attributes
are removed, whitespace compressed, and comments stripped. Windows-style
newlines are also converted to Unix-style terminators.

Finally, because this script was written to optimise the output for icon
SVGs only, embedded colour information is stripped to ensure geometry is
pure black (the default colour of SVG paths). This is done because black
is sometimes set to #181816 in Illustrator, as opposed to simply #000.

Usage: clean-icon-svg.pl [files...]