Add IE8 polyfills for Element-related properties
This code adds support in IE8 for the following properties:

  * Element.childElementCount
  * Element.firstElementChild
  * Element.lastElementChild
  * Element.nextElementSibling
  * Element.previousElementSibling

NOTE: IE8's implementation of Element.children is flawed in that Comment
nodes are included in the property's contents. These polyfills will work
around this shortcoming, but it's important to note that this code won't
address this behaviour.

The reason the polyfill doesn't fix this bug is because there's no clean
and easy way to do so. Simply overriding the property to return an array
of elements won't suffice; the result isn't a live HTMLCollection, so it
won't update with future DOM changes. To synthesise a dynamic collection
would require patching a *lot* of interfaces with hacky workarounds, and
for very little gain. If you need to support IE8 in your work, make sure
any elements whose children are being queried are absent of comments.