Add Git hook to prevent incomplete version-commits
If a commit's subject line matches a recognised version-commit ("Release
v1.0.1" or just "v1.0.1"), this hook will check package.json and any man
pages to make sure the version strings they contain are up-to-date.

If a discrepancy is encountered, the hook aborts the commit with a short
note explaining the oversight.