Update IE8-9 polyfills with Alhadis/Fix-IE@7b90e8a
A dedicated repo is now used to maintain these polyfills. This commit is
simply a result of locally hardlinking the files to keep them in sync.

Having said that, the "Fix-IE" repo should be consulted in future when a
polyfill is needed. The original copies will still be kept in this repo,
mainly to avoid breaking existing links (e.g.: tinyurl.com/fixIE8-9).